logic changes to setting isPublic
dataset:
if set to private -> dataset, datatables and datacols set to private
if set to public -> dataset set to public. datatables and data cols remain

datatable:
if set to private -> datatable and datacols set to private. dataset remains
if set to public -> dataset and datatable set to public. datacols remain

datacols:
if set to private -> datacols set to private. dataset and datatable remains
if set to public -> dataset, datatable and datacol set to public